requirements.txt
meteo-downloader
matplotlib~=3.10.6
pandas~=2.3.2
streamlit~=1.51.0
parse~=1.20.2
requests~=2.32.5
tabulate~=0.9.0
python-dotenv~=1.1.1
streamlit-authenticator~=0.4.2
pyyaml~=6.0.2